// frontend/src/client/types.gen.ts
export type AgentCreate = {
    name: string;
    description?: (string | null);
    role?: (string | null);
    goal?: (string | null);
    backstory?: (string | null);
    instructions?: (string | null);
};